backend/pyproject.toml
[project]
name = "amber-backend"
version = "0.1.0"
description = "FastAPI backend for Amber Badminton"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "mysql-connector-python",
    "trueskill",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "python-multipart",
    "pandas",
    "openpyxl",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "httpx",
]
